+++ actionpack/lib/action_dispatch/routing/mapper.rb
@@ -909,7 +909,7 @@ module ActionDispatch
         # CANONICAL_ACTIONS holds all actions that does not need a prefix or
         # a path appended since they fit properly in their scope level.
         VALID_ON_OPTIONS  = [:new, :collection, :member]
-        RESOURCE_OPTIONS  = [:as, :controller, :path, :only, :except, :param]
+        RESOURCE_OPTIONS  = [:as, :controller, :path, :only, :except, :param, :concerns]
         CANONICAL_ACTIONS = %w(index create new show update destroy)
 
         class Resource #:nodoc:
@@ -1045,6 +1045,8 @@ module ActionDispatch
 
           resource_scope(:resource, SingletonResource.new(resources.pop, options)) do
             yield if block_given?
+
+            concerns(options[:concerns]) if options[:concerns]
 
             collection do
               post :create
@@ -1209,6 +1211,8 @@ module ActionDispatch
 
           resource_scope(:resources, Resource.new(resources.pop, options)) do
             yield if block_given?
+
+            concerns(options[:concerns]) if options[:concerns]
 
             collection do
               get  :index if parent_resource.actions.include?(:index)
@@ -1580,15 +1584,71 @@ module ActionDispatch
           end
       end
 
+      # Routing Concerns allows you to declare common routes that can be reused
+      # inside others resources and routes.
+      #
+      #   concern :commentable do
+      #     resources :comments
+      #   end
+      #
+      #   concern :image_attachable do
+      #     resources :images, only: :index
+      #   end
+      #
+      # These concerns are used in Resources routing:
+      #
+      #   resources :messages, concerns: [:commentable, :image_attachable]
+      #
+      # or in a scope or namespace:
+      #
+      #   namespace :posts do
+      #     concerns :commentable
+      #   end
+      module Concerns
+        # Define a routing concern using a name.
+        #
+        #   concern :commentable do
+        #     resources :comments
+        #   end
+        #
+        # Any routing helpers can be used inside a concern.
+        def concern(name, &block)
+          @concerns[name] = block
+        end
+
+        # Use the named concerns
+        #
+        #   resources :posts do
+        #     concerns :commentable
+        #   end
+        #
+        # concerns also work in any routes helper that you want to use:
+        #
+        #   namespace :posts do
+        #     concerns :commentable
+        #   end
+        def concerns(*names)
+          names.flatten.each do |name|
+            if concern = @concerns[name]
+              instance_eval(&concern)
+            else
+              raise ArgumentError, "No concern named #{name} was found!"
+            end
+          end
+        end
+      end
+
       def initialize(set) #:nodoc:
         @set = set
         @scope = { :path_names => @set.resources_path_names }
+        @concerns = {}
       end
 
       include Base
       include HttpHelpers
       include Redirection
       include Scoping
+      include Concerns
       include Resources
     end
   end